backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
et_xmlfile==2.0.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
invoke==2.2.1
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
motor==3.7.1
netmiko==4.6.0
ntc_templates==8.1.0
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
paramiko==4.0.0
pluggy==1.6.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
PyNaCl==1.6.2
pyserial==3.5
pytest==9.0.2
pytest-sugar==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-nmap==0.7.1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
ruamel.yaml==0.19.1
scp==0.15.0
six==1.17.0
starlette==0.50.0
termcolor==3.3.0
textfsm==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
